+++ app/layout.js
@@ -1,7 +1,7 @@
 import { Banner } from '@/components/uswds/Banner';
 import { Header } from '@/components/Header';
-import { Identifier } from '@/components/Identifier';
-import { Footer } from '@/components/Footer';
+import { Identifier } from '@/components/uswds/Identifier';
+import { Footer } from '@/components/uswds/Footer';
 import { Sidebar } from '@/components/Sidebar';
 
 import '@/assets/stylesheets/styles.scss';
@@ -24,7 +24,7 @@ export default function RootLayout({ children }) {
           </div>
           <div className="desktop:grid-col-10 bg-white height-full flex-column display-flex">
             <Header />
-            <main className="padding-4 overflow-y-auto grow">{children}</main>
+            <main className="padding-4 overflow-y-auto">{children}</main>
           </div>
         </div>
         <Footer />
