+++ assets/stylesheets/styles.scss
@@ -5,15 +5,3 @@
 @forward 'uswds';
 
 // 3. Load any custom SASS
-.grow {
-  flex-grow: 1;
-}
-.grow-0 {
-  flex-grow: 0;
-}
-.shrink {
-  flex-shrink: 1;
-}
-.shrink-0 {
-  flex-shrink: 0;
-}
