+++ components/Header.tsx
@@ -7,7 +7,7 @@ export function Header() {
   const authSession = cookieStore.get('authsession');
 
   return (
-    <div className="desktop:display-flex flex-align-center flex-justify border-bottom border-accent-warm-light shrink">
+    <div className="desktop:display-flex flex-align-center flex-justify border-bottom border-accent-warm-light">
       <div className="padding-x-4 padding-y-2 desktop:order-last display-flex flex-justify-end border-bottom border-accent-warm-light desktop:border-0 font-ui-xs">
         {authSession ? <LogoutButton /> : <LoginButton />}
       </div>
